# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: GameHub API
  description: GameHub游戏社区平台API文档
  version: 1.0.0
  contact:
    name: GameHub开发团队
    email: dev@gamehub.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: 本地开发服务器
  - url: https://api.gamehub.example.com/api/v1
    description: 生产服务器
...
paths:
...
  /admin/analytics/user-growth:
    get:
      tags:
        - 管理员-分析
      summary: 用户增长统计
      description: 获取用户增长趋势数据
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly]
      responses:
        '200':
          description: 获取成功
  /admin/analytics/game-popularity:
    get:
      tags:
        - 管理员-分析
      summary: 游戏热度统计
      description: 获取游戏热度排行榜
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 获取成功
  /admin/analytics/content-engagement:
    get:
      tags:
        - 管理员-分析
      summary: 内容参与度统计
      description: 获取内容参与度分析
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 获取成功
  /admin/analytics/dashboard:
    get:
      tags:
        - 管理员-分析
      summary: 分析仪表盘
      description: 获取分析仪表盘数据
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 获取成功
...
  /admin/backups:
    get:
      tags:
        - 管理员-备份
      summary: 获取备份列表
      description: 获取所有备份记录
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 获取成功
    post:
      tags:
        - 管理员-备份
      summary: 创建备份
      description: 创建新的系统备份
      security:
        - bearerAuth: []
      responses:
        '201':
          description: 备份创建成功
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...